# goldrush/examples/stub_server/openapi_server/models/balance.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class Balance(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, balance=None, wallet=None):  # noqa: E501
        """Balance - a model defined in OpenAPI

        :param balance: The balance of this Balance.  # noqa: E501
        :type balance: int
        :param wallet: The wallet of this Balance.  # noqa: E501
        :type wallet: List[int]
        """
        self.openapi_types = {
            'balance': int,
            'wallet': List[int]
        }

        self.attribute_map = {
            'balance': 'balance',
            'wallet': 'wallet'
        }

        self._balance = balance
        self._wallet = wallet

    @classmethod
    def from_dict(cls, dikt) -> 'Balance':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The balance of this Balance.  # noqa: E501
        :rtype: Balance
        """
        return util.deserialize_model(dikt, cls)

    @property
    def balance(self):
        """Gets the balance of this Balance.


        :return: The balance of this Balance.
        :rtype: int
        """
        return self._balance

    @balance.setter
    def balance(self, balance):
        """Sets the balance of this Balance.


        :param balance: The balance of this Balance.
        :type balance: int
        """
        if balance is None:
            raise ValueError("Invalid value for `balance`, must not be `None`")  # noqa: E501

        self._balance = balance

    @property
    def wallet(self):
        """Gets the wallet of this Balance.

        Wallet with some coins.  # noqa: E501

        :return: The wallet of this Balance.
        :rtype: List[int]
        """
        return self._wallet

    @wallet.setter
    def wallet(self, wallet):
        """Sets the wallet of this Balance.

        Wallet with some coins.  # noqa: E501

        :param wallet: The wallet of this Balance.
        :type wallet: List[int]
        """
        if wallet is None:
            raise ValueError("Invalid value for `wallet`, must not be `None`")  # noqa: E501
        if wallet is not None and len(wallet) > 1000:
            raise ValueError("Invalid value for `wallet`, number of items must be less than or equal to `1000`")  # noqa: E501

        self._wallet = wallet