# goldrush/examples/stub_server/openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: HighLoad Cup 2021
  description: |
    ## Usage
    ## List of all custom errors
    First number is HTTP Status code, second is value of "code" field in returned JSON object, text description may or may not match "message" field in returned JSON object.
    - 422.1000: wrong coordinates
    - 422.1001: wrong depth
    - 409.1002: no more active licenses allowed
    - 409.1003: treasure is not digged
  version: 1.0.0
servers:
- url: /
paths:
...
  /balance:
    get:
      description: Returns a current balance.
      operationId: get_balance
      responses:
        "200":
          description: Current balance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/balance'
        default:
          description: General errors using same model as used by go-swagger for validation
            errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /licenses:
    get:
      description: Returns a list of issued licenses.
      operationId: list_licenses
      responses:
        "200":
          description: List of issued licenses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/licenseList'
        default:
          description: General errors using same model as used by go-swagger for validation
            errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      x-openapi-router-controller: openapi_server.controllers.default_controller
    post:
      description: Issue a new license.
      operationId: issue_license
      requestBody:
        description: Amount of money to spend for a license. Empty array for get free
          license. Maximum 10 active licenses
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/wallet'
        required: false
      responses:
        "200":
          description: Issued license.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/license'
        default:
          description: |
            - 409.1002: no more active licenses allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      x-codegen-request-body-name: args
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /explore:
    post:
      description: Returns amount of treasures in the provided area at full depth.
      operationId: explore_area
      requestBody:
        description: Area to be explored.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/area'
        required: true
      responses:
        "200":
          description: Report about found treasures.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/report'
        default:
          description: |
            - 422.1000: wrong coordinates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      x-codegen-request-body-name: args
      x-openapi-router-controller: openapi_server.controllers.default_controller
...
components:
  schemas:
    error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
          description: Either same as HTTP Status Code OR >= 600 with HTTP Status
            Code 422
          format: int32
        message:
          type: string
      description: |
        This model should match output of errors returned by go-swagger
        (like failed validation), to ensure our handlers use same format.
    balance:
      required:
      - balance
      - wallet
      type: object
      properties:
        balance:
          type: integer
          format: uint32
        wallet:
          $ref: '#/components/schemas/wallet'
      description: Current balance and wallet with up to 1000 coins.
      example:
        wallet:
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        - 6
        balance: 0
    wallet:
      maxItems: 1000
      uniqueItems: true
      type: array
      description: Wallet with some coins.
      items:
        type: integer
        format: uint32
    amount:
      minimum: 0
      type: integer
      description: Non-negative amount of treasures/etc.
    license:
      required:
      - digAllowed
      - digUsed
      - id
      type: object
      properties:
        id:
          type: integer
        digAllowed:
          $ref: '#/components/schemas/amount'
        digUsed:
          $ref: '#/components/schemas/amount'
      description: License for digging.
      example:
        digUsed: null
        digAllowed: 0
        id: 0
    licenseList:
      type: array
      description: List of issued licenses.
      items:
        $ref: '#/components/schemas/license'
    area:
      required:
      - posX
      - posY
      type: object
      properties:
        posX:
          minimum: 0
          type: integer
        posY:
          minimum: 0
          type: integer
        sizeX:
          minimum: 1
          type: integer
        sizeY:
          minimum: 1
          type: integer
      example:
        posX: 0
        posY: 0
        sizeX: 1
        sizeY: 1
    report:
      required:
      - amount
      - area
      type: object
      properties:
        area:
          $ref: '#/components/schemas/area'
        amount:
          $ref: '#/components/schemas/amount'
      example:
        area:
          posX: 0
          posY: 0
          sizeX: 1
          sizeY: 1
        amount: 0
...